+++ square.py
@@ -4,10 +4,10 @@ def area(a):
     Return area of square.
 
         Parameters:
-            a (double): Length of side of square.
+            a (float): Length of side of square.
         
         Return value:
-            area (double): Area of square.
+            area (float): Area of square.
     '''
     return a * a
 
@@ -17,9 +17,9 @@ def perimeter(a):
     Return perimeter of square.
 
         Parameters:
-            a (double): Length of side of square.
+            a (float): Length of side of square.
         
         Return value:
-            perimeter (double): Perimeter of square.
+            perimeter (float): Perimeter of square.
     '''
     return 4 * a
