+++ triangle.py
@@ -3,11 +3,11 @@ def area(a, h):
     Return area of triangle.
 
         Parameters:
-            a (double): Length of base of triangle.
-            h (double): Length of height of triangle.
+            a (float): Length of base of triangle.
+            h (float): Length of height of triangle.
         
         Return value:
-            area (double): Area of triangle.
+            area (float): Area of triangle.
     '''
     return a * h / 2 
 
@@ -16,11 +16,11 @@ def perimeter(a, b, c):
     Return area of triangle.
 
         Parameters:
-            a (double): Length of first side of triangle.
-            a (double): Length of second side of triangle.
-            a (double): Length of third side of triangle.
+            a (float): Length of first side of triangle.
+            a (float): Length of second side of triangle.
+            a (float): Length of third side of triangle.
         
         Return value:
-            perimeter (double): perimeter of triangle.
+            perimeter (float): perimeter of triangle.
     '''
     return a + b + c
